+++ WindowsFormsOrderingApp/DataAccessLayer/Entities/Stanje.cs
@@ -12,13 +12,13 @@ namespace DataAccessLayer.Entities
         public string Naziv { get; set; }
         public string JMJ { get; set; }
         public decimal Cijena { get; set; }
-        public decimal PSKolicina { get; set; }
+        public int PSKolicina { get; set; }
         public decimal PSIznos { get; set; }
-        public decimal UkupnaKolicinaUlaz { get; set; }
+        public int UkupnaKolicinaUlaz { get; set; }
         public decimal UkupniIznosUlaz { get; set; }
-        public decimal UkupnaKolicinaIzlaz { get; set; }
+        public int UkupnaKolicinaIzlaz { get; set; }
         public decimal UkupniIznosIzlaz { get; set; }
-        public decimal TrenutnoStanjeKolicina { get; set; }
+        public int TrenutnoStanjeKolicina { get; set; }
         public decimal TrenutnoStanjeIznos { get; set; }
     }
 }
